Support implicit conversion from a type value to a facet value
If a type satisfies the requirements of a FacetType, then that type
as a value can be converted to a FacetValue, which binds the type value
to the FacetType.

For instance, if the class A implements an interface B, then

  fn F(b:! B) {}

can be called with the type `A`

  F(A);

This does not handle yet receiving non-type values matching a FacetType,
as that requires deducing the required FacetValue for the caller's
argument. Follow-up work will do this step.