Deduce facet values for arguments to generic fns receiving a facet type
The fn receiving a facet type needs to deduce a type from a
FacetAccessType, which is the SemIR type representing the parameter type
that is a generic parameter. For example:

```
  fn F[T: Interface](val: T);
```

Here T is a generic parameter that is a facet value, but the `val`
parameter's type is the facet value converted from a FacetType to a
TypeType, with `as type`. The result of that conversion is a
FacetAccessType. So the deduction code sees a FacetAccessType for the
type of `val`.

We make deduction undo the `as type` conversion to move back to the `T`
parameter declaration, which has type FacetType in order to deduce the
required facet value (which is itself a type constrained by the
FacetType). And when we have a facet value (of type FacetType) to be
deduced, we will also convert the argument if it is of an appropriate
type to a FacetValue that matches the FacetType using the changes from
PR #4863.

This exposes an infinite recursion in impl lookup when deducing a
FacetValue for an interface that is implemented via `impl forall`. We
resolve that by adding a shortcut in impl lookup, where if the impl
itself is concrete then we check that the ids match and do not do
deduction on the implicit parameters of the `impl forall` statement
(which would end up looking back at the `impl forall`).

If the impl constraint is symbolic, the same recursive loop should still
be possible to hit, but that does not typecheck through deduction yet.
There are some fail_todo tests for deducing and converting from a
facet-value-value (non-type value that is symbolic and contrained by a
facet type).

Rebased on top of PR #4885