Handle FacetAccessWitness in eval of ImplWitnessAccess
ImplWitnessAccess evaluation was expecting the impl witness instruction
to always be an ImplWitness, but the witness can also be a
FacetAccessWitness, which needs to be looked through to find the
ImplWitness.

Adds a test that would CHECK fail without this change. It still fails
typecheck but without a crash, and is left with a TODO. The test is
derived from
toolchain/check/fuzzer_corpus/ca69be54b8273f334ab8c7db0ad454c5ff0291c9.

Closes #4868