Improve comments on conversion from type value to type id
The GetTypeIdForTypeConstant() and GetTypeIdForTypeInst() methods
convert from a type value to the type id of that value. Expand the
comments to explain why they can not be called with facet values, which
would be of type `FacetType`.

Change the CHECK inside to disallow being called with facet values as
well. No tests cause this to occur anymore.