Add enumerate() for ValueStore, ConstantValueStore, SpecificStore, ImplStore
This allows iterating on all values in a store along with the Id for
each value, instead of `llvm::enumerate(store.array_ref())` which would
give you the indices.

While the indices are really the same as the Ids, this provides a
typesafe way to enumerate() over a store.

For ConstantValueStore the indices are InstIds. This replaced the one
use of array_ref() so it is removed.

Replace loops over llvm::seq() of the size of a store, or
llvm::enumerate() of the array of a store, with a call to .enumerate().