Reverse nesting of BoundMethod and SpecificFunction.
During SemIR, when identifying a specific in a specific context, we'd
have to either look through a specific or through a bound method. Canonicalize
which one to look through first, by having the BoundMethod created around a
SpecificFunction instead.
This changes a lot of check tests.

TODO: As the SemIr does not currently allow removal, the bound
instruction created prior to finding the specific is not removed from
the instructions.
Options: (1) leave as is, (2) add a way to remove teh previous bound,
(3) rethink how/when the BoundMethod inst is created.