Add PRELUDE: option for file tests to specify a minimal prelude library
The PRELUDE option is only used in the toolchain tests for now. If
specified in a `no_prelude` test, the given file path is used as a
replacement for the production prelude library. The path is specified
relative to the repository root.

To support this:
* //testing knows how to parse PRELUDE out of the header of a test file.
* //testing gives this to the FileTest implementation through a new
  virtual method.
* //toolchain/testing received the path to the minimal prelude file and
  uses it to tell the driver the path to use as the prelude.
* //toolchain/driver grows the --prelude-path-for-testing command line
  flag to specify where the prelude should be loaded from in these
  tests.
* //toolchain/testing/min_prelude/ is a directory of carbon files that
  can be used as a prelude. They are not include in the `all_testdata`
  filegroup to avoid running them as tests, but are placed into a new
  `minimal_preludes` filegroup so they are installed and can be accessed
  during test execution.

Tests that were creating their own minimal prelude to define BitAnd on
types are now pointed to toolchain/testing/min_prelude/facet_types.carbon
as the prelude. They no longer need to `import Core` in each test as a
result.