Support conversion from facet value to facet value
When converting from a facet value (an instruction whose type is
FacetType), we require making a FacetAccessType
to have a type instruction when building the resulting FacetValue.
Otherwise, the conversion is the same for values of type TypeType, and
we relax the convert function to support either.

There is a special case of doing impl lookup directly on the FacetType
for a facet value, in order to find impls of an interface on an
interface (impl Iface1 as Iface2), and this is kept separate, and moved
to lower priority than trying to convert the facet value itself. We note
explicitly that this only works for FacetTypes of a single interface, as
we don't expect `A & B` to satisfy an `impl A as Y` and be treated as
`A & B` having an impl for `Y`.

We support impl lookup on FacetAccessType to find interfaces satisifed
by it, in order to support conversion from symbolic values of type
FacetAccessType (FacetType as type) to a new symbolic value of a
different FacetType. Deduce moves from converting the values to
converting the types, when the types are symbolic, which makes it
attempt to convert FacetAccessType to a facet value (ie. a value of
type FacetType). Then the impl lookup has to be done through the
facet type (of the FacetAccessType) as we don't know the concrete type
in order to do impl lookup on it while inside a generic function. This
enables the following tests to pass by supporting the conversion of `w`,
which has type FacetAccessType:
```
 impl forall [A:! Animal] A as Eats {}

 fn FeedTame2[V:! Tame & Eats](v: V) {}

 fn HandleTameAnimal2[W:! Animal & Tame](w: W) {
   FeedTame2(w);
 }
```