Build the full type structure for an impl
Keep traversing type structure after finding first symbolic, and add
concrete/symbolic items to a structural representation in a vector. When
we find a type with testing, we mark the opening/closing of that type so
that a symbolic will know which things it matches against. If it's
positioned at the same place as an concrete opening, then it matches
everything up to and including the paired concrete closing.

Matching with this structure is not implemented yet. Everything just
returns true.