Remove position_ in type structure builder
The position is the size of the structure, now that we are building
the full structure. To help make this all more clear we move the
pushing into the type structure vector into a helper method and
have it record the position of the first symbolic appended to the
type structure.

We use 'append' for the operation of adding to the type structure
to try differentiate it from 'push' which is used to add work to
the work stack.