Diagnose the presence of generic params on a concrete impl decl
These parameters are never deduced, so they prevent the impl decl from
ever being used. But we also don't emit diagnostics inside impl lookup,
so there's nothing provided to the user explaining that they made an
impl that is useless.

For now we just diagnose the case where no generic parameters are used.
It would be nice to also diagnose when some but not all are used.